mock_server/requirements.txt
Flask==2.2.3
requests==2.28.2
Werkzeug==2.3.7
timm==0.9.2
torch
Pillow
python-dotenv==1.0.0
logging==0.4.9.6
numpy
pydicom==2.3.1
transformers